\name{describeSensor-methods}

\docType{methods}

\alias{describeSensor}
\alias{describeSensor-methods}
\alias{describeSensor,SOS,character-method}
\alias{describeSensor,SOS_1.0.0,character-method}

\title{Method for a DescribeSensor Request to a SOS}

\description{This method sends a DescribeSensor request for a description of the given procedure to the given Sensor Observation Service instance.}

\section{Methods}{
	\describe{
		\item{\code{signature(sos = "SOS", procedure = "character")}}{Method requests a description of the given procedure from the given SOS.}
	}
}

\keyword{methods}
