\name{getCapabilities-methods}

\docType{methods}

\alias{getCapabilities}
\alias{getCapabilities-methods}
\alias{getCapabilities,SOS-method}

\title{Request Capabilities from a SOS}

\description{
This method request the metadata description of a given Sensor Observation Service, the Capabilities document.
}

\section{Methods}{

\describe{
	\item{\code{signature(sos = "SOS")}}{Request capabilities description from a SOS.}
}

}

\keyword{methods}