\name{KML}
\Rdversion{1.1}

\alias{parseKML}
\alias{mimeTypeKML}
\alias{kmlName}
\alias{kml}


\title{Methods for the Namespace kml}

\description{
Methods for handling responses in Keyhole Markup Language (KML) format.

At the current stage, the parsers simply returns an object from the package \pkg{XML} with the KML document.
}

\references{
\url{https://www.opengeospatial.org/standards/kml}
\url{https://developers.google.com/kml/documentation/}
}

\author{Daniel Nuest \email{daniel.nuest@uni-muenster.de}}

\examples{
#
}

\keyword{methods}
\keyword{misc}
