\name{sosObservableProperties-methods}

\docType{methods}

\alias{sosObservableProperties}
\alias{sosObservableProperties-methods}
\alias{sosObservableProperties,list-method}
\alias{sosObservableProperties,OmObservation-method}
\alias{sosObservableProperties,OmObservationCollection-method}
\alias{sosObservableProperties,SOS-method}
\alias{sosObservableProperties,SosObservationOffering_2.0.0-method}
\alias{sosObservableProperties,SweCompositePhenomenon-method}
\alias{sosObservableProperties,SwePhenomenonProperty-method}

\title{Methods for Function \code{sosObservableProperties} in Package \pkg{sos4R}}

\description{Methods for function \code{sosObservableProperties} in package \pkg{sos4R}. The function allows to retrieve observable properties, e.g. air temperature, wind speed, etc., listed in the capabilities of a Sensor Observation Service.}
\section{Methods}{
\describe{

\item{\code{signature(obj = "list")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(obj = "OmObservation")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(obj = "OmObservationCollection")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(obj = "SOS")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(obj = "SosObservationOffering_2.0.0")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(obj = "SweCompositePhenomenon")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(obj = "SwePhenomenonProperty")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
