\name{SweTextEncoding-class}

\Rdversion{1.1}

\docType{class}

\alias{SweTextEncoding-class}
\alias{SweTextEncoding}

\title{Class \code{"SweTextEncoding"}}
\description{
Representation of a text encoding defined in the OGC SWE Common specification. It defines the token, block, and decimal seperators for a text-encoded array of values.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("SweTextEncoding", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{tokenSeparator}:}{Object of class \code{"character"} ~~ }
    \item{\code{blockSeparator}:}{Object of class \code{"character"} ~~ }
    \item{\code{decimalSeparator}:}{Object of class \code{"character"} ~~ }
    \item{\code{id}:}{Object of class \code{"character"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "SweTextEncoding" in the signature.
}
\references{
%%  ~~put references to the literature/web site here~~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("SweTextEncoding")
}
\keyword{classes}
