\name{FoiOrNULL-class}
\Rdversion{1.1}
\docType{class}

\alias{FoiOrNULL-class}

\title{Class \code{"FoiOrNULL"}}
\description{
Classes and construction functions for features of interest from the OGC Observations and Measurements specification. A feature of interest is considered to be the real world feature that is observed.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "FoiOrNULL" in the signature.
}

\examples{
showClass("FoiOrNULL")
}
\keyword{classes}
