\name{getCapabilities-methods}

\docType{methods}

\alias{getCapabilities}
\alias{getCapabilities-methods}
\alias{getCapabilities,SOS-method}
\alias{getCapabilities,SOS_1.0.0-method}
\alias{getCapabilities,SOS_1.0.0,verbose,inspect-method}

\title{Request Capabilities from a SOS}

\description{
This method request the metadata description of a given Sensor Observation Service, the Capabilities document.
}

\section{Methods}{

\describe{
	\item{\code{signature(sos = "SOS")}}{Request capabilities description from a SOS.}
	\item{\code{signature(inspect = "logical")}}{Print out the sent and received documents to console.}
	\item{\code{signature(verbose = "logical")}}{Extensive debugging information printed to console.}
}

}

\keyword{methods}
