\name{KML}
\Rdversion{1.1}

\alias{parseKML}
\alias{mimeTypeKML}
\alias{kmlName}
\alias{kml}


\title{Methods for the Namespace kml}

\description{
Methods for handling responses in Keyhole Markup Language (KML) format.

At the current stage, the parsers simply returns an object from the package \pkg{XML} with the KML document.
}

\references{
\url{http://www.opengeospatial.org/standards/kml}

\url{http://code.google.com/intl/de-DE/apis/kml/documentation/}
}

\author{Daniel Nüst \email{daniel.nuest@uni-muenster.de}}

\examples{
#
}

\keyword{methods}
\keyword{misc}
